Remove the useless ShowRepository interface
It is now an implementation class, like all the other
repositories
